Create HTML
 Accessing the HTML content from webpage 

Let us try to understand this piece of code.

First of all import the requests library.

Then, specify the URL of the webpage you want to scrape.
Send a HTTP request to the specified URL and save the response from server in a response object called r.

Now, as print r.content to get the raw HTML content of the webpage. It is of ‘string’ type.

Note: Sometimes you may get error “Not accepted” so try adding a browser user agent like below. Find your user agent based on device and browser from here

 https://deviceatlas.com/blog/list-of-user-agent-strings